working routers for plaid / query controller
